+++ src/Modules/Basket/BasketView.php
@@ -174,12 +174,6 @@ class BasketView extends View
 
 	public function basketForm($foodsaver)
 	{
-		global $g_data;
-		$g_data['weight'] = '3';
-		$g_data['contact_type'] = 1;
-		$g_data['tel'] = $foodsaver['telefon'];
-		$g_data['handy'] = $foodsaver['handy'];
-
 		$out = '';
 
 		$out .= $this->v_utils->v_form_textarea('description', array('maxlength' => 1705));
@@ -206,18 +200,20 @@ class BasketView extends View
 		}
 
 		$out .= $this->v_utils->v_form_select('weight', array(
-			'values' => $values
+			'values' => $values,
+			'selected' => 3
 		));
 
 		$out .= $this->v_utils->v_form_checkbox('contact_type', array(
 			'values' => array(
 				array('id' => 1, 'name' => 'Per Nachricht'),
 				array('id' => 2, 'name' => 'Per Telefonanruf')
-			)
+			),
+			'checked' => [1]
 		));
 
-		$out .= $this->v_utils->v_form_text('tel');
-		$out .= $this->v_utils->v_form_text('handy');
+		$out .= $this->v_utils->v_form_text('tel', ['value' => $foodsaver['telefon']]);
+		$out .= $this->v_utils->v_form_text('handy', ['value' => $foodsaver['handy']]);
 
 		$out .= $this->v_utils->v_form_checkbox('food_type', array(
 			'values' => array(