+++ src/Modules/Core/Control.php
@@ -5,10 +5,13 @@ namespace Foodsharing\Modules\Core;
 use Foodsharing\DI;
 use Foodsharing\Lib\Db\Mem;
 use Foodsharing\Lib\Func;
+use Foodsharing\Lib\Sanitizer;
 use Foodsharing\Lib\Session\S;
+use Foodsharing\Lib\Twig;
 use Foodsharing\Lib\View\Utils;
 use Foodsharing\Modules\Message\MessageModel;
 use ReflectionClass;
+use Symfony\Component\HttpFoundation\Request;
 
 abstract class Control
 {
@@ -29,6 +32,11 @@ abstract class Control
 	 * @var Utils
 	 */
 	protected $v_utils;
+
+	/**
+	 * @var Twig
+	 */
+	private $twig;
 
 	public function __construct()
 	{
@@ -82,6 +90,22 @@ abstract class Control
 			}
 		}
 		$this->model->updateActivity(S::id());
+	}
+
+	/**
+	 * @required
+	 */
+	public function setTwig(Twig $twig)
+	{
+		$this->twig = $twig;
+	}
+
+	protected function render($template, $data)
+	{
+		$global = $this->func->generateAndGetGlobalViewData();
+		$viewData = array_merge($global, $data);
+
+		return $this->twig->render($template, $viewData);
 	}
 
 	public function setTemplate($template)
@@ -493,5 +517,27 @@ abstract class Control
 		}
 
 		return false;
+	}
+
+	protected function sanitizeRequest(Request $request, $spec)
+	{
+		$data = [];
+		foreach ($spec as $name => $s) {
+			$default = ['method' => 'get', 'required' => true, 'parameterName' => $name, 'default' => null];
+			$s = array_merge($default, $s);
+			$v = $request->request->{$s['method']}($s['parameterName']);
+			if (is_null($v)) {
+				if ($s['required']) {
+					throw new \Exception('Required parameter not set');
+				}
+				$v = $s['default'];
+			}
+			if (isset($s['filter'])) {
+				$v = call_user_func([Sanitizer::class, $s['filter']], $v);
+			}
+			$data[$name] = $v;
+		}
+
+		return $data;
 	}
 }