+++ src/Modules/Core/View.php
@@ -394,15 +394,9 @@ class View
 	{
 		$this->func->addHead('<script src="https://maps.googleapis.com/maps/api/js?v=3.exp&libraries=places&key=' . GOOGLE_API_KEY . '"></script>');
 
-		global $g_data;
-		if (isset($g_data['lat']) && isset($g_data['lon']) && !empty($g_data['lat']) && !empty($g_data['lon'])) {
-			$data = array(
-				'lat' => $g_data['lat'],
-				'lon' => $g_data['lon'],
-				'zoom' => 14
-			);
+		if (isset($options['location'])) {
+			$data = array_merge(['zoom' => 14], $options['location']);
 		} else {
-			/* Todo: Remove when getting rid of $g_data */
 			$db = DI::$shared->get(Model::class);
 			$data = $db->getValues(array('lat', 'lon'), 'foodsaver', $this->func->fsId());
 			$data['zoom'] = 14;
@@ -443,30 +437,19 @@ class View
 				$(\'#lon\').val(result.lng());
 				$(\'#plz\').val(result.nameForType(\'postal_code\'));
 				$(\'#ort\').val(result.nameForType(\'locality\'));
-				if($(\'#anschrift\').length) {
-					$(\'#anschrift\').val(address + (number ? (\' \' + number):\'\')) 
-				}
-				else {
-					$(\'#str\').val(address)
-					$(\'#hsnr\').val(number)
-				}
+				$(\'#anschrift\').val(address + (number ? (\' \' + number):\'\'));
 			});
 			$("#lat-wrapper,#lon-wrapper").hide();
 		');
 
-		$hsnr = $this->v_utils->v_form_text('anschrift', array('disabled' => '1', 'required' => '1'));
-		if (isset($options['hsnr'])) {
-			$hsnr = $this->v_utils->v_form_text('str', array('required' => '1')) . $this->v_utils->v_form_text('hsnr');
-		}
-
 		return $this->v_utils->v_input_wrapper($this->func->s('position_search'), '
 		<input placeholder="Straße, Ort..." type="text" value="" id="addresspicker" type="text" class="input text value ui-corner-top" />
 		<div id="map" class="pickermap"></div>') .
-			$hsnr .
-			$this->v_utils->v_form_text('plz', array('disabled' => '1', 'required' => '1')) .
-			$this->v_utils->v_form_text('ort', array('disabled' => '1', 'required' => '1')) .
-			$this->v_utils->v_form_text('lat') .
-			$this->v_utils->v_form_text('lon') .
+			$this->v_utils->v_form_text('anschrift', ['value' => $options['anschrift'], 'disabled' => '1', 'required' => '1']) .
+			$this->v_utils->v_form_text('plz', ['value' => $options['plz'], 'disabled' => '1', 'required' => '1']) .
+			$this->v_utils->v_form_text('ort', ['value' => $options['ort'], 'disabled' => '1', 'required' => '1']) .
+			$this->v_utils->v_form_text('lat', ['value' => $options['lat']]) .
+			$this->v_utils->v_form_text('lon', ['value' => $options['lon']]) .
 			'';
 	}