+++ tests/_support/Helper/Foodsharing.php
@@ -276,7 +276,7 @@ class Foodsharing extends \Codeception\Module\Db
 		return $this->createRegion($name, $parentId, 7, $extra_params);
 	}
 
-	public function createRegion($name, $parentId = 741, $type = 8, $extra_params = [])
+	public function createRegion($name, $parentId = 741, $type = 9, $extra_params = [])
 	{
 		/* 741 is germany, so suitable for normal sub regions */
 		$v = array_merge([