+++ tests/_support/Helper/FoodsharingUI.php
@@ -1,10 +1,4 @@
 <?php
-/**
- * Created by IntelliJ IDEA.
- * User: matthias
- * Date: 13.02.18
- * Time: 13:13.
- */
 
 namespace Helper;
 
@@ -23,16 +17,21 @@ class FoodsharingUI extends \Codeception\Module
 	 */
 	public function addInTagSelect($value, $tagEditInSelector, $tagEditId = 'tagedit')
 	{
-		$inputId = '#' . $tagEditId . '-input';
+		$inputId = $tagEditInSelector . ' #' . $tagEditId . '-input';
 		$this->getBrowser()->clickWithLeftButton($tagEditInSelector, 3, 3);
 		$this->getBrowser()->fillField($inputId, $value);
 		$selector = '//a[contains(@id, \'ui-id\') and contains(text(), "' . $value . '")]';
 		$this->getBrowser()->click($selector);
 	}
 
-	public function removeFromTagSelect($value)
+	public function removeFromTagSelect($value, $tagEditInId = null)
 	{
-		$selector = '//*[@value="' . $value . '"]/following-sibling::*';
+		if ($tagEditInId) {
+			$selector = '//*[@id="' . $tagEditInId . '"]//*[@value="' . $value . '"]/following-sibling::*';
+		} else {
+			$selector = '//*[@value="' . $value . '"]/following-sibling::*';
+		}
+
 		$this->getBrowser()->click($selector);
 		/* wait until it is gone, it might change the layout */
 		$this->getBrowser()->dontSee($selector);