+++ tests/_support/Helper/FoodsharingUrl.php
@@ -57,4 +57,24 @@ class FoodsharingUrl extends \Codeception\Module\Db
 	{
 		return '/?page=fairteiler&bid=' . (int)$region_id;
 	}
+
+	public function fairTeilerGetUrlShort($fairteiler_id)
+	{
+		return '/fairteiler/' . (int)$fairteiler_id;
+	}
+
+	public function fairTeilerGetUrl($fairteiler_id)
+	{
+		return '/?page=fairteiler&sub=ft&id=' . (int)$fairteiler_id;
+	}
+
+	public function fairTeilerEditUrl($fairteiler_id)
+	{
+		return '/?page=fairteiler&sub=ft&id=' . (int)$fairteiler_id . '&sub=edit';
+	}
+
+	public function foodBasketInfoUrl($basket_id)
+	{
+		return '/essenskoerbe/' . (int)$basket_id;
+	}
 }